clean up folder organization
