+++ src/robot_module/setup.py
@@ -4,7 +4,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 setup_args = generate_distutils_setup(
-     packages=['robot_module'],
+     packages=['robot_module', 'locks', 'services'],
      package_dir={'': 'src'}
 )
