--- src/robot_module/src/robot_module/__init__.py
@@ -1,3 +1,0 @@
-from .main import RobotModule
-
-__all__ = ["RobotModule"]
